requirements.txt
streamlit>=1.32.0
plotly>=5.20.0
pandas>=2.0.0
numpy>=1.26.0
requests>=2.31.0
groq>=0.8.0
python-dotenv>=1.0.0
supabase>=2.0.0
